[WIP] add recipe overview to nav and move views to ui package
